requirements.txt
numpy
scipy
matplotlib
tk
py-pde